feat: support CREATE TRIGGER/FUNCTION/EVENT/PROCEDURE task check
