+++ plugin/parser/utils.go
@@ -1,7 +1,9 @@
 package parser
 
 import (
+	"fmt"
 	"io"
+	"regexp"
 	"strings"
 
 	tidbast "github.com/pingcap/tidb/parser/ast"
@@ -66,4 +68,61 @@ func SetLineForMySQLCreateTableStmt(node *tidbast.CreateTableStmt) error {
 	}
 	firstLine := node.OriginTextPosition() - strings.Count(node.Text(), "\n")
 	return newTokenizer(node.Text()).setLineForMySQLCreateTableStmt(node, firstLine)
+}
+
+// ExtractTiDBUnsupportStmts returns a list of unsupported statements in TiDB extracted from the `stmts`,
+// and returns the remaining statements supported by TiDB from `stmts`.
+func ExtractTiDBUnsupportStmts(stmts string) ([]string, string, error) {
+	var (
+		unsupportStmts []string
+		supportedStmts string
+	)
+	// We use our bb tokenizer to help us split the multi-statements into statement list.
+	singleSQLs, err := SplitMultiSQL(MySQL, stmts)
+	if err != nil {
+		return nil, "", errors.Wrapf(err, "cannot split multi sql %q via bytebase parser", stmts)
+	}
+	for _, sql := range singleSQLs {
+		content := sql.Text
+		if isTiDBUnsupportStmt(content) {
+			unsupportStmts = append(unsupportStmts, content)
+			continue
+		}
+		supportedStmts += content + "\n"
+	}
+	return unsupportStmts, supportedStmts, nil
+}
+
+// isTiDBUnsupportStmt returns true if this statement is unsupported in TiDB.
+func isTiDBUnsupportStmt(stmt string) bool {
+	if isTiDBUnsupportDDLStmt(stmt) {
+		return true
+	}
+	// Match DELIMITER statement
+	// Now, we assume that all input comes from our mysqldump, and the tokenizer can split the mysqldump DELIMITER statement
+	// in one singleSQL correctly, so we can handle it easily here by checking the prefix.
+	return strings.HasPrefix(stmt, "DELIMITER ")
+}
+
+// isTiDBUnsupportStmt checks whether the `stmt` is unsupported DDL statement in TiDB, the following statements are unsupported:
+// 1. `CREATE TRIGGER`
+// 2. `CREATE EVENT`
+// 3. `CREATE FUNCTION`
+// 4. `CREATE PROCEDURE`.
+func isTiDBUnsupportDDLStmt(stmt string) bool {
+	objects := []string{
+		"TRIGGER",
+		"EVENT",
+		"FUNCTION",
+		"PROCEDURE",
+	}
+	regexFmt := "(?i)^CREATE DEFINER=`.+`@`.+` %s\\s+"
+	for _, obj := range objects {
+		regex := fmt.Sprintf(regexFmt, obj)
+		re := regexp.MustCompile(regex)
+		if re.MatchString(stmt) {
+			return true
+		}
+	}
+	return false
 }
